css/css-values/minmax-length-percent-serialize.html

<!DOCTYPE html>
<link rel="help" href="https://drafts.csswg.org/css-values-4/#comp-func">
<link rel="help" href="https://drafts.csswg.org/css-values-4/#mixed-percentages">
<link rel="help" href="https://drafts.csswg.org/css-values-4/#calc-serialize">
<link rel="author" title="Xiaocheng Hu" href="mailto:xiaochengh@chromium.org">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../support/parsing-testcommon.js"></script>
<script>

function test_computed(value, expected) {
    test_valid_value('flex-basis', value, expected);
}

// If fully resolvable to a number, serialize to a calc().
test_computed('min(1px)', 'calc(1px)');
test_computed('max(1px)', 'calc(1px)');

// If not, keep as the function.
test_computed('min(1% + 1px)', 'min(1% + 1px)');
test_computed('min(1px + 1%)', 'min(1% + 1px)');
test_computed('max(1px + 1%)', 'max(1% + 1px)');

// Arguments are simplified, but not reordered.
test_computed('min(20px, 10%)', 'min(20px, 10%)');
test_computed('min(1em, 10%)', 'min(16px, 10%)');
test_computed('min(10%, 20px)', 'min(10%, 20px)');
test_computed('min(10%, 1em)', 'min(10%, 16px)');
test_computed('max(20px, 10%)', 'max(20px, 10%)');
test_computed('max(1em, 10%)', 'max(16px, 10%)');
test_computed('max(10%, 20px)', 'max(10%, 20px)');
test_computed('max(10%, 1em)', 'max(10%, 16px)');

// Within an argument, normal sorting occurs
test_computed('min(10% + 30px, 60px + 5%)', 'min(10% + 30px, 5% + 60px)');
test_computed('max(10% + 2em, 1em + 5%)', 'max(10% + 32px, 5% + 16px)');

// min()/max() are valid inside a calc(),
// and retain their relative order
test_computed('calc(min(10% + 1px) + max(10% + 1em) + min(10% + 20px))', 'calc(min(10% + 1px) + max(10% + 16px) + min(10% + 20px))');

// min()/max() can be combined with plain units as well.
// While min()/max() maintain their own ordering,
// ordinary units will re-sort around them.
test_computed('calc(1em + max(10% + 20px) + 5% + min(10% + 1em) + 10px)', 'calc(5% + 26px + max(10% + 20px) + min(10% + 16px))');
</script>
